% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_haver_codes.R
\name{parse_haver_codes}
\alias{parse_haver_codes}
\title{Wrapper function around \code{Haver::haver.data} that returns a tibble with the dates in Date format and each series as a column.}
\usage{
parse_haver_codes(series)
}
\arguments{
\item{series}{The series (in database:series or SERIES@DATABASE format) that you wish to pull from Haver.}
}
\value{
A character vector of Haver codes in the format database:
}
\description{
Wrapper function around \code{Haver::haver.data} that returns a tibble with the dates in Date format and each series as a column.
}
\examples{
parse_haver_codes('GDP@USECON')
}
