% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_haver.R
\name{import_haver}
\alias{import_haver}
\title{Wrapper function around \code{Haver::haver.data} that returns a tibble with
the dates in Date format and each series as a column.}
\usage{
import_haver(series, eop = TRUE, ...)
}
\arguments{
\item{series}{The series (in database:series or SERIES@DATABASE format) that
you wish to pull from Haver. Optionally, you can pass series names that
differ from the Haver mnemonics as names for the character vector for Haver 
codes.}

\item{eop}{Logical (TRUE/FALSE) whether to use end of period dates for annual,
quarterly, and monthly series. Defaults to TRUE.}

\item{...}{Additional arguments to be passed to \code{haver.data}. The most 
useful are \code{start} and \code{end}.}
}
\value{
A \code{tibble} with a date column and a column for each series.
}
\description{
Wrapper function around \code{Haver::haver.data} that returns a tibble with
the dates in Date format and each series as a column.
}
\examples{
import_haver(c("usecon:gdp", "usecon:c"))
}
