% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_haver_codes.R
\name{parse_haver_codes}
\alias{parse_haver_codes}
\title{This function parses haver codes of the format
code{SERIES@DATABASE} and converts them to the format \code{database:series}}
\usage{
parse_haver_codes(series)
}
\arguments{
\item{series}{A series name to parse into the proper format}
}
\value{
A character vector of Haver codes in the format "database:series"
}
\description{
This function parses haver codes of the format
code{SERIES@DATABASE} and converts them to the format \code{database:series}
}
\examples{
parse_haver_codes("GDP@USECON")
}
